backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
sqlmodel>=0.0.16
psycopg2-binary>=2.9.5
python-jose[cryptography]>=3.3.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
asyncpg>=0.27.0
passlib>=1.7.4
bcrypt>=4.0.0